+++ resources/views/requests/index.blade.php
@@ -48,7 +48,9 @@
                     </a>
                 </td>
                 <td class="text-center">
-                    {{ $request->summary->queries_count }}
+                    <a href="/sense/queries?filter[request][id]={{ $request->id }}">
+                        {{ $request->summary->queries_count }}
+                    </a>
                 </td>
                 <td class="text-center">
                     {{ $request->summary->time_total }}